201804201211
Added a new extension: buildingrequest. This is a new instance for our housing, building and development department.
Added a new documentfile for this profile
Added a new profile for this instance
Added translationfiles.

This is based on the rcx request profile and instance, this is still a WIP, still have to customize everything to the needs of this department. First test.