201807231639
added in BuildingRequest a new sort of tickets: BuildingViolation.
This is meant to be able to have a better follow-up on all building and environmental violations. It has a lot of custom fields, specifically for violations.
Added translations for both English and Dutch.
This is the rough (working) version, it probably still needs some tweaking.