201807301453
changed the way BuildingRequest, BuildingViolation and RcxRequest can be re-assigned. Instead of directly in the ticket, it has to be with the re-assign button.
For this, I added re-assignrights to the user profiles and changed the attributes to read_only.
Also added read and write rights to the Building User for BuildingViolation.