201807311543
made a new ticketing system: HR Request. This will be used by our Human Resources department to follow everything up on sicknesses, staff turnover,... over multilple physical sites.
This is completely with a new menu at the left and with a new user profile. Only people with the HR User Profile will be able to see this (confidentiality is important here!)
At this moment it's just the rough version, but it works. Still have to customize it to their wishes and needs.