201904091532
added a new class "Participation". The purpose of this new class is to be able to let citizens give suggestions to our city, where other citizens can then vote on if they think it's a good idea.
As of this moment, this is just the suggestions part. The vote part still has to be written.